assignment_4/requirements.txt
torch == 2.2.2
pillow == 10.2.0
facenet-pytorch == 2.6.0
pandas == 2.2.2
matplotlib == 3.9.0 
seaborn == 0.13.2
tqdm == 4.66.4